+++ src/SM.h
@@ -34,7 +34,7 @@ void StationMonitor() {
   // =================================================================
   memset(msgbuf, 0, sizeof(msgbuf));
   sprintf (msgbuf+strlen(msgbuf), "DM:%d DR:%d %s", 
-    distance_gauge_median(), analogRead(DISTANCEGAUGE),
+    distance_gauge_median(), (int) analogRead(DISTANCEGAUGE),
     (dg_adjustment == 1.25) ? "5M" : "10M");
 
   len = (strlen (msgbuf) > 21) ? 21 : strlen (msgbuf);
